services/dbt-service/requirements.txt
dbt-core==1.2.1
dbt-bigquery==1.2.0
Flask>=1.1.3
gunicorn==20.1.0
PyYAML==6.0